requirements.txt
beautifulsoup4==4.6.1
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
dj-static==0.0.6
Django==1.11.15
django-modelcluster==4.1
django-overextends==0.4.3
django-taggit==0.22.2
django-toolbelt==0.0.1
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.0.0
gunicorn==19.9.0
html5lib==0.999999999
idna==2.7
Pillow==5.2.0
psycopg2==2.7.5
pytz==2018.5
requests==2.19.1
six==1.11.0
static3==0.7.0
Unidecode==0.4.21
urllib3==1.23
wagtail==2.0.1
webencodings==0.5.1
whitenoise==3.3.1
Willow==1.1
